create tupaia server package
