+++ packages/tupaia-server/README.md
@@ -1,0 +1,3 @@
+# @tupaia/lesmi-server
+
+Back end server for the LESMIS web app.
