+++ packages/tupaia-server/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@tupaia/lesmis-server",
-  "version": "0.0.0",
+  "name": "@tupaia/tupaia-server",
+  "version": "1.0.0",
   "private": true,
-  "description": "Backend server for LESMIS application",
+  "description": "Backend server for boilerplate",
   "homepage": "https://github.com/beyondessential/tupaia",
   "bugs": {
     "url": "https://github.com/beyondessential/tupaia/issues"
@@ -21,24 +21,14 @@
     "prestart": "npm run -s build",
     "start": "node dist",
     "start-dev": "LOG_LEVEL=debug ../../scripts/bash/backendStartDev.sh 9998 -ts",
-    "start-verbose": "LOG_LEVEL=debug npm run start-dev",
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "test:watch": "jest --watch"
+    "start-verbose": "LOG_LEVEL=debug npm run start-dev"
   },
   "dependencies": {
     "@babel/polyfill": "^7.0.0",
     "@tupaia/access-policy": "3.0.0",
     "@tupaia/database": "1.0.0",
-    "@tupaia/tupaia-server": "1.0.0",
     "@tupaia/utils": "1.0.0",
-    "api-error-handler": "^1.0.0",
-    "body-parser": "^1.18.3",
-    "camelcase-keys": "^6.2.2",
-    "client-sessions": "^0.8.0",
-    "cors": "^2.8.5",
     "express": "^4.16.2",
-    "jsonwebtoken": "^8.5.1",
-    "winston": "^3.2.1"
+    "jsonwebtoken": "^8.5.1"
   }
 }
