+++ packages/tupaia-server/src/ApiConnection.ts
@@ -1,9 +1,10 @@
-/**
+/*
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
  */
 import { fetchWithTimeout, verifyResponseStatus, stringifyQuery } from '@tupaia/utils';
-import { FetchConfig, QueryParameters, RequestBody } from '../types';
+import { FetchConfig, QueryParameters, RequestBody } from './types';
 
 interface AuthHandler {
   email?: string;
