+++ packages/tupaia-server/src/AuthConnection.ts
@@ -1,9 +1,10 @@
-/**
+/*
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
  */
 
-import { AuthResponse, Credentials } from '../types';
+import { AuthResponse, Credentials } from './types';
 import { ApiConnection } from './ApiConnection';
 
 const {
