+++ packages/tupaia-server/src/types.ts
@@ -1,11 +1,12 @@
-/**
+/*
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
  */
 
 import { Request } from 'express';
 import { ParamsDictionary, Query } from 'express-serve-static-core';
-import { LesmisSessionModel, LesmisSessionType } from './models';
+import { SessionModel, SessionType } from './Session';
 
 export type AccessPolicyObject = Record<string, string[]>;
 
@@ -23,9 +24,9 @@ export interface LesmisRequest<>extends Request<
     LesmisRequestBody,
     Query
   > {
-  sessionModel: LesmisSessionModel;
+  sessionModel: SessionModel;
   sessionCookie?: SessionCookie;
-  session?: LesmisSessionType;
+  session?: SessionType;
 }
 
 export interface LesmisResponseBody {
@@ -63,4 +64,20 @@ export interface AuthResponse {
   accessToken?: string;
   refreshToken?: string;
   user?: AuthResponseUser;
+}
+
+export interface SessionDetails {
+  email: string;
+  accessPolicy: AccessPolicyObject;
+  accessToken: string;
+  refreshToken: string;
+}
+
+export interface SessionFields {
+  id: string;
+  email: string;
+  access_policy: AccessPolicyObject;
+  access_token: string;
+  access_token_expiry: number;
+  refresh_token: string;
 }
